requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.5
Django==4.2.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
exceptiongroup==1.1.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
PyJWT==2.7.0
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytz==2023.3
redis==4.5.5
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
psycopg2
